+++ CHANGELOG.md
@@ -5,6 +5,20 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [6.1.0] 2025-08-01
+
+### Fixed
+
+- ML_optimization: Make correct use of the evotorch library.
+
+### Changed
+
+- ML_inference: `PyTorchInferenceHandler` will use 4 methods to handle return types as numpy array or torch tensor:
+    - `predict_numpy()`
+    - `predict_batch_numpy()`
+    - `predict()`
+    - `predict_batch()`
+
 ## [6.0.1] 2025-08-01
 
 ### Fixed